requirements.txt
asgiref==3.7.2
Django==4.2.10
djangorestframework==3.15.1
sqlparse==0.4.4
tzdata==2024.1
